Add files via upload
fmdserver - This file will create a virtual server which will transfer data between different users who are online. It will maintain all the created Threads in a sorted manner and will identify the threads in which data has to be transferred when any user will enter any message. This file is responsible for identifying Group chat message and  Personal Chat message.

fmdclient - This file is user creation file. This file have to executed one time per user means  for each user this file has to be executed once. When fmdclient  will be executed it will ask for username which will act as a username for the user till he logout. For Logout user simply can type bye